RDNA2: fix RQ/TQ bf16 dtype mismatch + attention tuning + import optimization
Three improvements for RDNA2 (gfx1030/gfx1031) inference:

1. Fix RotorQuant/TurboQuant bf16 dtype mismatch (root cause)
   - model_runner_kv_cache_mixin.py: KV cache pools were hardcoded to
     torch.bfloat16 regardless of model compute dtype. On RDNA2 where
     bf16 is unsupported and dtype is forced to fp16, this caused
     tl.dot(Q, K) mixed-dtype Triton errors in AITER's unified_attention.
   - Changed dtype=torch.bfloat16 -> dtype=self.dtype at all 3 pool
     creation sites (MLA TQ, MHA TQ, MHA RQ).

2. RDNA2-tuned attention kernel configs
   - decode_attention.py: BLOCK=32 + num_warps=2 for RDNA2 Wave32
     (upstream uses BLOCK=8 + num_warps=1 as MI3xx SGPR workaround,
     massively underutilizing RDNA2 LDS/wavefronts).
   - extend_attention.py: num_warps=2 for RDNA2 (upstream uses 4),
     bf16->fp16 dtype casts for RDNA2 compatibility, restored k_extend
     dimension handling.

3. Lift RDNA2 imports out of forward_hip() hot loops
   - activation.py: rdna2_ops, rdna2_silu_and_mul, rdna2_gelu_and_mul
     moved to module-level try/except ImportError. Also added missing
     'import os' for os.environ usage in _check_rdna2_activations().
   - layernorm.py: rdna2_ops, rdna2_fused_add_rms_norm, rdna2_rms_norm
     moved to module-level try/except ImportError.
   - rotary_embedding/base.py: rdna2_ops moved to module-level
     try/except ImportError.
   Python's import checks sys.modules on every call (~1-5us). In
   forward methods called thousands of times/sec, this is measurable.

+++ python/sglang/srt/model_executor/model_runner_kv_cache_mixin.py
@@ -599,7 +599,7 @@ class ModelRunnerKVCacheMixin:
                 self.token_to_kv_pool = MLATokenToKVPoolTQ(
                     self.max_total_num_tokens,
                     page_size=self.page_size,
-                    dtype=torch.bfloat16,
+                    dtype=self.dtype,
                     kv_lora_rank=self.model_config.kv_lora_rank,
                     qk_rope_head_dim=self.model_config.qk_rope_head_dim,
                     layer_num=self.num_effective_layers,
@@ -732,7 +732,7 @@ class ModelRunnerKVCacheMixin:
                     self.token_to_kv_pool = MHATokenToKVPoolTQ(
                         self.max_total_num_tokens,
                         page_size=self.page_size,
-                        dtype=torch.bfloat16,
+                        dtype=self.dtype,
                         head_num=self.model_config.get_num_kv_heads(
                             get_attention_tp_size()
                         ),
@@ -758,7 +758,7 @@ class ModelRunnerKVCacheMixin:
                     self.token_to_kv_pool = MHATokenToKVPoolRQ(
                         self.max_total_num_tokens,
                         page_size=self.page_size,
-                        dtype=torch.bfloat16,
+                        dtype=self.dtype,
                         head_num=self.model_config.get_num_kv_heads(
                             get_attention_tp_size()
                         ),
